# [ATen][CUDA] Implement 128 bit vectorization v2