@Aidyn-A 2025-01-27
> Did we confirm not regresion on A100 per comments on the prev PR?

Yes, the regression is due to a compiler (NVCC) bug. Moreover, I discovered the bug is present on H100 as well. I have omitted vec8/vec16 for 1-byte data on all archs to workaround the bug.