@Aidyn-A 2025-01-27
> What versions of NVCC are affected? We are potentially planning dropping old NVCC support (11.8/12.4)

I have not checked 11.8, but 12.4-12.6 are affected. 12.8 is doing fine if `nvvm-latest` flag was enforced. Though, I am not certain if it is generally safe to force `nvvm-latest` on sm_80-90.