@ZainRizvi 2025-01-29
@pytorchbot revert -c nosignal -m "Sorry but this is breaking in trunk. See functorch/test_ops.py::TestOperatorsCUDA::test_jvp_nn_functional_multi_head_attention_forward_cuda_float32 [GH job link](https://github.com/pytorch/pytorch/actions/runs/13032483748/job/36358184032) [HUD commit link](https://hud.pytorch.org/pytorch/pytorch/commit/81685d81eb86595d169f55a564da26eaafb2ddf5)"