@eqy 2025-01-30
I have a feeling this failure is going to be maxwell-specific...