@Aidyn-A 2025-01-30
> I have a feeling this failure is going to be maxwell-specific...

I am seeing it on A100 as well. How did CI miss it in the first go? :thinking: 
Anyways, I am debugging it.