@ngimel 2025-01-30
We expect this PR to be bitwise-identical to previous implementation (right?), so any change in values seems real.