@Aidyn-A 2025-01-31
> We expect this PR to be bitwise-identical to previous implementation (right?), so any change in values seems real. 

It is bitwise-identical. I was referring to benchmark of compilation time.