@ngimel 2025-02-25
I think it helps perf even on H100