@malfet 2025-02-25
> I think it helps perf even on H100

I've made off by one error in my calculations anyway, CUDA_ARCH > 10 are Fermi's and above :)